LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-history-replaceState-expected.txt


FAIL event.intercept() should proceed if the given promise resolves assert_equals: expected "#1" but got ""
